Further update deps and move a function from common to utils
