+++ packages/cli/package.json
@@ -15,6 +15,7 @@
     "boxen": "5.1.2",
     "cli-ux": "^6.0.9",
     "ejs": "^3.1.10",
+    "fs-extra": "^11.2.0",
     "fuzzy": "^0.1.3",
     "json5": "^2.2.3",
     "node-fetch": "2.7.0",
@@ -42,6 +43,7 @@
     "@subql/common-stellar": "^4.0.0",
     "@subql/common-substrate": "workspace:*",
     "@types/ejs": "^3.1.5",
+    "@types/fs-extra": "^11.0.4",
     "@types/node": "^18.19.42",
     "@types/semver": "^7.5.8",
     "@types/update-notifier": "^6",
