+++ packages/common/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- `findAvailablePort` function, it now lives in `@subql/utils` (#2518)
 
 ## [5.0.0] - 2024-08-01
 ### Added
