+++ packages/common/package.json
@@ -18,17 +18,12 @@
     "axios": "^0.28.0",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.1",
-    "detect-port": "^1.6.1",
-    "fs-extra": "^11.2.0",
     "js-yaml": "^4.1.0",
     "reflect-metadata": "^0.1.14",
     "semver": "^7.6.3",
     "update-notifier": "7.2.0"
   },
   "devDependencies": {
-    "@types/bn.js": "5.1.5",
-    "@types/detect-port": "^1.3.5",
-    "@types/fs-extra": "^11.0.4",
     "@types/js-yaml": "^4.0.9",
     "@types/semver": "^7.5.8",
     "@types/update-notifier": "^6"
