+++ packages/common/src/project/readers/ipfs-reader.ts
@@ -4,7 +4,7 @@
 import {u8aConcat} from '@polkadot/util';
 import {Reader} from '@subql/types-core';
 import yaml from 'js-yaml';
-import {IPackageJson} from 'package-json-type';
+import type {IPackageJson} from 'package-json-type';
 import {IPFS_NODE_ENDPOINT} from '../../constants';
 import {IPFSHTTPClientLite} from '../IpfsHttpClientLite';
 
@@ -17,10 +17,7 @@ export class IPFSReader implements Reader {
   private ipfs: IPFSHTTPClientLite;
   private cache: Record<string, Promise<string>> = {};
 
-  constructor(
-    private readonly cid: string,
-    gateway?: string
-  ) {
+  constructor(private readonly cid: string, gateway?: string) {
     if (!CIDv0.test(cid) && !CIDv1.test(cid)) {
       throw new Error('IPFS project path CID is not valid');
     }
