+++ packages/types-core/src/project/readers/reader.ts
@@ -1,7 +1,7 @@
 // Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
 // SPDX-License-Identifier: GPL-3.0
 
-import {IPackageJson} from 'package-json-type';
+import type {IPackageJson} from 'package-json-type';
 
 export type ReaderOptions = {
   ipfs?: string;
