+++ packages/utils/package.json
@@ -16,18 +16,16 @@
     "@polkadot/util": "^13.0.2",
     "@polkadot/util-crypto": "^13.0.2",
     "@subql/x-sequelize": "6.32.0-0.0.4",
-    "ansi-styles": "^6.2.1",
     "chalk": "^4.1.2",
+    "detect-port": "^1.6.1",
     "flatted": "^3.3.1",
     "graphql": "^15.8.0",
     "graphql-tag": "^2.12.6",
     "lodash": "^4.17.21",
     "pino": "^9.3.2",
-    "rotating-file-stream": "^3.2.3",
-    "semver": "^7.6.3",
-    "tar": "^7.4.3"
+    "rotating-file-stream": "^3.2.3"
   },
   "devDependencies": {
-    "@types/semver": "^7.5.8"
+    "@types/detect-port": "^1.3.5"
   }
 }
