+++ packages/utils/src/networking.ts
@@ -1,0 +1,19 @@
+// Copyright 2020-2024 SubQuery Pte Ltd authors & contributors
+// SPDX-License-Identifier: GPL-3.0
+
+import detectPort from 'detect-port';
+
+export async function findAvailablePort(startPort: number, range = 10): Promise<number | null> {
+  for (let port = startPort; port <= startPort + range; port++) {
+    try {
+      const _port = await detectPort(port);
+      if (_port === port) {
+        return port;
+      }
+    } catch (e) {
+      return null;
+    }
+  }
+
+  return null;
+}
